Mostly Tests
* More tests
* Added cleanup code to a couple existing tests
* Improved the updateDbVersion() test. Previously, it created a file of
the latest version and ran the method. The method only checked the
version, saw that it was the latest, and exited the function. Now, it
creates the database as it was when the version updates were first
implemented before running the method. This causes it to test every
update script.
* Fixed a bug in CardManager.giveAllTypesTitleFields()